+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.3
+
+- Add a method to get a rune word from the dictionary.
+- Add methods to update the common dictionary and knowledge map.
+- Update the dictionary and knowledge map when received on the client.
+
 ## 1.0.2
 
 - Add a debug message to show that the player has learned a word.